3d_api/requirements_conda.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.2
appdirs==1.4.4
apptools==5.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
body-measurements==1.1.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.3
coloredlogs==15.0.1
configobj==5.0.6
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.20
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
envisage==6.1.0
executing==1.1.1
fastapi==0.87.0
fastjsonschema==2.16.2
filelock==3.8.0
filetype==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==22.9.24
flatten-dict==0.4.2
fonttools==4.37.4
freetype-py==2.3.0
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gast==0.4.0
gdown==4.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
huggingface-hub==0.11.1
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.22.2
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipdb==0.13.11
ipykernel==5.3.4
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jpeg4py==0.1.4
json5==0.9.10
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyter_packaging==0.12.3
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.1
kaolin @ git+https://github.com/NVIDIAGameWorks/kaolin.git@bcce7a95506847b3ced48fcfc3207f61488dcb42
keras==2.11.0
kiwisolver==1.4.4
kornia==0.6.8
lap==0.4.0
libclang==14.0.6
llvmlite==0.39.1
loguru==0.5.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mayavi==4.8.1
mediapipe==0.8.11
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.2
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.5.1
notebook_shim==0.2.0
numba==0.56.3
numpy==1.22.4
oauthlib==3.2.2
onnxruntime==1.13.1
open3d==0.15.2
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plotly==5.11.0
ply==3.11
pooch==1.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyface==7.4.4
pyglet==1.5.27
Pygments==2.13.0
PyMatting==1.1.8
PyMCubes==0.1.2
pymeshlab==2022.2.post2
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.5
pyparsing==3.0.9
PyQt5==5.14.2
PyQt5-sip==12.11.0
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.2.5
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@276c9a8acbfa715f5802f26ec9f4141bde26ecb4
pytz==2022.5
PyWavefront==1.3.3
PyWavelets==1.4.1
pyzmq==24.0.1
rembg @ git+https://github.com/YuliangXiu/rembg.git@5f4386b2f1e9464ea7138e225fabf819e9ee42d1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rtree==0.9.7
scikit-image==0.19.1
scikit-learn==1.1.2
scipy==1.7.2
Send2Trash==1.8.0
setuptools-scm==7.0.5
Shapely==1.7.1
simple-romp==1.0.4
sip==6.7.5
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
starlette==0.21.0
sympy==1.11.1
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.4.12
tinycss2==1.2.1
tinyobjloader==2.0.0rc7
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
torchmetrics==0.10.0
tornado==6.2
tqdm==4.62.3
traitlets==5.5.0
traits==6.4.1
traitsui==7.4.2
trimesh==3.9.35
urllib3==1.26.12
usd-core==22.5.post1
uvicorn==0.20.0
vedo==2020.4.2
voxelize-cuda @ git+https://github.com/YuliangXiu/neural_voxelization_layer.git@b3c054e378fcca4daf6a53aceeee7ef39e86318b
vtk==9.2.2
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wrapt==1.14.1
wslink==1.8.4
yarl==1.8.1
zipp==3.9.0
